// server/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ChatType = {
  DIRECT: 'DIRECT',
  GROUP: 'GROUP'
} as const

export type ChatType = (typeof ChatType)[keyof typeof ChatType]


export const MemberRole = {
  ADMIN: 'ADMIN',
  MEMBER: 'MEMBER'
} as const

export type MemberRole = (typeof MemberRole)[keyof typeof MemberRole]


export const MessageType = {
  TEXT: 'TEXT',
  IMAGE: 'IMAGE',
  VIDEO: 'VIDEO',
  AUDIO: 'AUDIO',
  VOICE: 'VOICE',
  DOCUMENT: 'DOCUMENT',
  CALL: 'CALL'
} as const

export type MessageType = (typeof MessageType)[keyof typeof MessageType]


export const MessageStatus = {
  SENT: 'SENT',
  DELIVERED: 'DELIVERED',
  READ: 'READ'
} as const

export type MessageStatus = (typeof MessageStatus)[keyof typeof MessageStatus]


export const CallType = {
  AUDIO: 'AUDIO',
  VIDEO: 'VIDEO'
} as const

export type CallType = (typeof CallType)[keyof typeof CallType]


export const CallStatus = {
  ONGOING: 'ONGOING',
  COMPLETED: 'COMPLETED',
  MISSED: 'MISSED',
  REJECTED: 'REJECTED',
  CANCELED: 'CANCELED'
} as const

export type CallStatus = (typeof CallStatus)[keyof typeof CallStatus]